backend/requirements/base.txt
asgiref
Django
django-filter
djangorestframework
Markdown
django-environ==0.4.5  # https://github.com/joke2k/django-environ
djangorestframework_simplejwt